+++ packages/models/lib/main.reflectable.dart
@@ -2,21 +2,8 @@
 // https://github.com/dart-lang/reflectable.
 
 import 'dart:core';
-
 import 'package:device_frame/src/info/device_type.dart' as prefix33;
 import 'package:meta/meta.dart' as prefix31;
-// ignore_for_file: camel_case_types
-// ignore_for_file: implementation_imports
-// ignore_for_file: prefer_adjacent_string_concatenation
-// ignore_for_file: prefer_collection_literals
-// ignore_for_file: unnecessary_const
-
-// ignore:unused_import
-import 'package:reflectable/mirrors.dart' as m;
-// ignore:unused_import
-import 'package:reflectable/reflectable.dart' as r show Reflectable;
-// ignore:unused_import
-import 'package:reflectable/src/reflectable_builder_based.dart' as r;
 import 'package:theta_models/src/widgets/dynamic_attributes_parse.dart'
     as prefix0;
 import 'package:theta_models/src/widgets/features/align.dart' as prefix1;
@@ -64,6 +51,19 @@ import 'package:theta_models/src/widgets/intrinsic_states/intrinsic_state_parse.
     as prefix29;
 import 'package:theta_models/src/widgets/nodes/key_constants.dart' as prefix32;
 import 'package:theta_models/src/widgets/nodes/node_type.dart' as prefix34;
+
+// ignore_for_file: camel_case_types
+// ignore_for_file: implementation_imports
+// ignore_for_file: prefer_adjacent_string_concatenation
+// ignore_for_file: prefer_collection_literals
+// ignore_for_file: unnecessary_const
+
+// ignore:unused_import
+import 'package:reflectable/mirrors.dart' as m;
+// ignore:unused_import
+import 'package:reflectable/src/reflectable_builder_based.dart' as r;
+// ignore:unused_import
+import 'package:reflectable/reflectable.dart' as r show Reflectable;
 
 final _data = <r.Reflectable, r.ReflectorData>{
   const prefix0.DynamicAttributeKey(): r.ReflectorData(