+++ packages/models/lib/src/widgets/nodes/nnull.dart
@@ -17,6 +17,7 @@ class NNull extends CNode {
           adapter: const NNullWidgetAdapter(),
           updatedAt: DateTime.now(),
           pageID: '',
+          isLocked: false,
         );
 
   static NNull fromJson(
@@ -54,6 +55,7 @@ class NNull extends CNode {
     NodeID? stabilID,
     PageID? componentID,
     List<CNode>? componentChildren,
+    bool? isLocked,
   }) {
     throw UnimplementedError('copyWith not implemented in NNull class');
   }
@@ -74,6 +76,7 @@ class NNull extends CNode {
     NodeID? stabilID,
     PageID? componentID,
     List<CNode>? componentChildren,
+    bool? isLocked,
   }) {
     throw UnimplementedError('copyWithWithNull not implemented in NNull class');
   }