+++ packages/widgets/CHANGELOG.md
@@ -1,3 +1,16 @@
+## 0.2.2
+
+- [x] Feat: add support to models 0.2.0.
+- [x] Fix: unused widgets removed.
+
+## 0.2.1
+
+- [x] Fix: image_network package removed.
+
+## 0.2.0
+
+- [x] Fix: Cors proxy removed from image.
+
 ## 0.1.9
 
 - [x] Feat: Visibility on laptops added.